Move TestOpener to common
